+++ src/main/java/com/flab/eattofit/member/application/auth/OAuthJsonMapper.java
@@ -1,4 +1,4 @@
-package com.flab.eattofit.member.domain.auth;
+package com.flab.eattofit.member.application.auth;
 
 import com.flab.eattofit.member.infrastructure.auth.dto.OAuthUserResponse;
 import com.flab.eattofit.member.infrastructure.auth.dto.UserInfoKeyWordRequest;
