+++ src/main/java/com/flab/eattofit/member/infrastructure/auth/OAuthManagerImpl.java
@@ -1,8 +1,8 @@
 package com.flab.eattofit.member.infrastructure.auth;
 
 import com.flab.eattofit.member.application.auth.OAuthManager;
-import com.flab.eattofit.member.domain.auth.OAuthConnectionManager;
-import com.flab.eattofit.member.domain.auth.OAuthJsonMapper;
+import com.flab.eattofit.member.application.auth.OAuthConnectionManager;
+import com.flab.eattofit.member.application.auth.OAuthJsonMapper;
 import com.flab.eattofit.member.infrastructure.auth.dto.OAuthProviderRequest;
 import com.flab.eattofit.member.infrastructure.auth.dto.OAuthUserResponse;
 import lombok.RequiredArgsConstructor;
